Source delete: Be sure to check if the source has a labelset.

+++ project/images/views.py
@@ -333,7 +333,8 @@ def source_admin(request, source_id):
             labelset = source.labelset
 
             source.delete()
-            labelset.delete()
+            if labelset:
+                labelset.delete()
 
             messages.success(request, 'Source has been deleted.')
